Change the way in which materials are loaded
Materials now check to make sure that any of their meltable items exist before creation. So if nothing can be used to make a material, it won't get loaded.

This is a good change because before, we grouped materials with mods, so if a mod was present, the materials associated with that mod would load. This was changed because a material could potentially be associated with items from multiple mods.